+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Bound `AbstractPostEntityMapper` to container to ease process of creating new
   post types and to reuse the same instance across all post types
 
+### Added
+- Added `url` method to `Comment` model
+
 ## [2.1.1] - 2015-10-27
 ### Fixed
 - Fixing a backwards incompatible change introduced in 2.1.0 where